+++ 1P2020_201442819/src/Interprete/Expresiones/TipoPrimitivo.java
@@ -12,7 +12,7 @@ package Interprete.Expresiones;
 public enum TipoPrimitivo {
     LIST,
     STRING,
-    DOUBLE,
+    NUMERIC,
     INTEGER,
     BOOL
 }
